+++ src/mrs/static/manifest.json
@@ -1,6 +1,6 @@
 {
   "short_name": "MRS",
-  "name": "Mes Remboursements Simplifies",
+  "name": "Mes Remboursements Simplifiés",
   "icons": [
     {
       "src": "launcher-icon-1x.png",
